Moved median functions to medianController
